pyproject.toml
[build-system]
requires = ["uv_build >= 0.8.17, <0.9.0"]
build-backend = "uv_build"

[project]
name = "rubric"
version = "1.1.7"
description = "rubric"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Gavin", email = "gavin@llmdata.com" },
    { name = "Joey", email = "joey@llmdata.com" }
]
keywords = ["rubrics", "grading", "eval", "rubric"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[project.urls]
Homepage = "https://github.com/The-LLM-Data-Company/rubrics/"
Repository = "https://github.com/The-LLM-Data-Company/rubrics"
Issues = "https://github.com/The-LLM-Data-Company/rubrics/issues"

